+++ NEWS.md
@@ -1,6 +1,7 @@
 # dials 0.0.1.9000
 
 * `regularization` was changed to `penalty` in a few models to be consistent with [this change](tidymodels/model-implementation-principles@08d3afd). 
+* Parameter objects now contain code to finalize their values and a number of helper functions for cetain data-specific parameters. 
 
 
 # dials 0.0.1